pyproject.toml
[tool.poetry]
name = "dsmusic"
version = "0.6.7"
description = "A simple music bot"
license = "MIT"
authors = ["John Toniutti <john.toniutti@gmail.com>"]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: aiohttp",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3"
aiohttp = "^3.9"
orjson = "^3.9"
aiodns = "^3.1"
brotli = "^1.1"
mafic = "^2.10"
setuptools = "^69"


[tool.poetry.group.test.dependencies]
flake8 = "^7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
